restructure folders + adds perf charts + placement changes
