+++ src/app/Analytics/LocalMonitoring/RecommendationsForLocalMonitoring/RemoteMonitoring/RecommendationComponents/LinePlot/ChartDataPreparation.tsx
