+++ src/app/Analytics/LocalMonitoring/RecommendationsForLocalMonitoring/RemoteMonitoring/RecommendationComponents/LinePlot/PerfHistoricCharts.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { Chart, ChartAxis, ChartGroup, ChartLine, ChartVoronoiContainer } from '@patternfly/react-charts';
-import { formatTimestamps, filterDataByTerm, formatNumber } from './ChatDataPreparation';
-import { Grid, GridItem, Split, SplitItem } from '@patternfly/react-core';
+import { formatTimestamps, filterDataByTerm, formatNumber } from './ChartDataPreparation';
+import { Grid, GridItem} from '@patternfly/react-core';
 
-const CostHistoricCharts = (props: { chartData; day; endtime }) => {
+const PerfHistoricCharts = (props: { chartData; day; endtime }) => {
   const termFilteredData = filterDataByTerm(props.chartData, props.endtime, props.day);
   const timeStampFormattedData = formatTimestamps(termFilteredData);
 
   const cpuChart = () => {
     const historicdata = Object.entries(timeStampFormattedData).map(([key, value]) => {
       const cpuAmount = formatNumber(
-        (value as any).recommendation_terms[props.day]?.recommendation_engines?.cost?.config?.requests?.cpu?.amount
+        (value as any).recommendation_terms[props.day]?.recommendation_engines?.performance?.config?.requests?.cpu?.amount
       );
       const displayKey = props.day === 'short_term' ? key.split(' ')[1] : key.split(' ')[0];
       return {
@@ -37,7 +37,7 @@ const CostHistoricCharts = (props: { chartData; day; endtime }) => {
           legendOrientation="vertical"
           legendPosition="right"
           height={250}
-          name="Cost CPU Recommendations"
+          name="Performance CPU Recommendations"
           domainPadding={{ y: [30, 25], x: [30, 25] }}
           padding={{
             bottom: 70,
@@ -80,7 +80,7 @@ const CostHistoricCharts = (props: { chartData; day; endtime }) => {
   const memoryChart = () => {
     const historicdata = Object.entries(timeStampFormattedData).map(([key, value]) => {
       const memoryAmount = formatNumber(
-        (value as any).recommendation_terms[props.day]?.recommendation_engines?.cost?.config?.requests?.memory?.amount
+        (value as any).recommendation_terms[props.day]?.recommendation_engines?.performance?.config?.requests?.memory?.amount
       );
       const displayKey = props.day === 'short_term' ? key.split(' ')[1] : key.split(' ')[0];
       return {
@@ -107,7 +107,7 @@ const CostHistoricCharts = (props: { chartData; day; endtime }) => {
           legendOrientation="vertical"
           legendPosition="right"
           height={250}
-          name="Cost Memory Recommendations"
+          name="Performance Memory Recommendations"
           domainPadding={{ y: [30, 25], x: [30, 25] }}
           padding={{
             bottom: 70,
@@ -160,4 +160,4 @@ const CostHistoricCharts = (props: { chartData; day; endtime }) => {
   );
 };
 
-export { CostHistoricCharts };
+export { PerfHistoricCharts };
